Add the fake package to %INC for newer import::into
As of version 1.002003 of import::into, it uses Module::Runtime to load
the package being imported into automatically. This breaks our use case
of using "fake" packages which we are creating on the fly.

By adding the fake package's notional filename to %INC manually we can
bypass this from the outside. It's a bit hacky, but then what is
Test::Kit if not a bunch of hacks cobbled together and hidden away
behind a nicer API? :-)