BIG CHANGE - Add support for Test::Builder >= 1.3
As of Test::Builder 1.3 the Test::Builder::Module package for creating
things which both export functions and behave like Test::More is
deprecated in favour of a slightly different mechanism provided by
Test::Builder::Provider.

In general Test::Builder::Provider works more nicely than
Test::Builder::Module, but because we want to maintain backwards
compatibility with older versions of Test::Builder we have to jump
through some hoops. See the comments in the code for more details.

On the plus side while working on this I also stumbled across a solution
to the "it's impossible to recognise exported scalars". Using
Scalar::Util::refaddr() it actually is possible to check that the symbol
table entry seen refers to an exported scalar, if you know what package
it was exported to. Nice!